Split entity platform from entity component
